支持配置 obfuscationLevel和 obfuscationPercentage,支持advancedObfuscation::neg

+++ Editor/Utils/RandomWithKey.cs
@@ -50,7 +50,7 @@
 
         public float NextFloat()
         {
-            return (float)((double)NextInt() / int.MaxValue);
+            return (float)((double)(uint)NextInt() / uint.MaxValue);
         }
 
         public bool NextInPercentage(float percentage)
